syncwell_rag_backend/requirements.txt
fastapi
uvicorn
sentence-transformers
google-generativeai
python-dotenv
chromadb